Don't check process if pidfile is empty or doesn't exist

+++ lib/spork/task_helper.rb
@@ -121,6 +121,7 @@ module Spork
     end
 
     def process_running?
+      return false unless pid
       begin
         Process.kill(0, pid)
         return true
